[refactor]: 라우터 모듈 분리 - /chat/room/:roomId/leave - /chat/room/:roomId
